+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Fix a bug giving an javascript error when you creating a new empty formula
 - Fix a bug preventing DB backup when running in cloud or on prem
 - Auto redirect back to the login page when session is expired
+- Make PV Online import dialog box persistent untile the import is complete
+- Set size for ingredients search box to auto to comply with screen size
 
 ### Version 6.1
 - Stock management improvements